Honor sync cancellation checks
Extract generated runtime driver

Route sync lowering through node lowerers

Extract flow expression classification
